+++ src/transparent-qgw-database/api-validation.test.ts
@@ -1,4 +1,4 @@
-import { apiSchema } from "./api-validation";
+import { apiSchema } from "./api/validation";
 
 // Zod helpers for clean debugging
 function expectZodSuccess(parsed: ReturnType<typeof apiSchema.safeParse>) {
