+++ src/transparent-qgw-database/api/validation.ts
