+++ src/transparent-qgw-database/transaction-request/options.ts
@@ -1,4 +1,4 @@
-import { DirectAPI, TransactionType } from "..";
+import { DirectAPI, TransactionType } from "../api";
 import {
   toggle1or2,
   toggleYesOrNO,
