+++ src/transparent-qgw-database/transaction-request/payment.ts
@@ -1,4 +1,4 @@
-import { DirectAPI } from "..";
+import { DirectAPI } from "../api";
 
 // cvvType: CVVtype "0" | "1" | "2" | "9" === Not Passing CVV2 | CVV2 is being passed | CVV2 on card is unreadable |Card does not have CVV2 imprint
 export enum CvvType {
